rename useGraphql to be subgraph specific
