+++ components/ExtraneousWarning.tsx
@@ -3,7 +3,7 @@ import { useAccount } from "wagmi";
 import { Alert, Button, Container } from "@mui/material";
 
 import { getDaoManagerQuery } from "@graphql/queries/subgraph/get-dao-manager-query";
-import { useGraphQL } from "@graphql/useGraphql";
+import { useSubgraphGraphQL } from "@graphql/subgraph";
 
 import useLogout from "@hooks/useLogout";
 import useResolutionsAcl from "@hooks/useResolutionsAcl";
@@ -11,7 +11,7 @@ import useResolutionsAcl from "@hooks/useResolutionsAcl";
 export default function ExtraneousWarning({ children }: { children: React.ReactNode }) {
   const { acl, isLoading } = useResolutionsAcl();
   const { isConnected, address, isConnecting } = useAccount();
-  const { data, isLoading: isLoadingGetDaoManager, error } = useGraphQL(address ? getDaoManagerQuery : null);
+  const { data, isLoading: isLoadingGetDaoManager, error } = useSubgraphGraphQL(address ? getDaoManagerQuery : null);
   const { logout } = useLogout();
 
   if ((!isLoadingGetDaoManager && data?.daoManager === null) || error) {
