+++ components/dashboard/Tokens.tsx
@@ -3,7 +3,7 @@ import useSWR from "swr";
 import { Box, Divider, Typography } from "@mui/material";
 
 import { getDaoManagerQuery } from "@graphql/queries/subgraph/get-dao-manager-query";
-import { useGraphQL } from "@graphql/useGraphql";
+import { useSubgraphGraphQL } from "@graphql/subgraph";
 
 import UserBalance from "@components/tokens/UserBalance";
 
@@ -11,7 +11,7 @@ import useUserBalanceAndOffers, { bigIntToNum } from "@hooks/useUserBalanceAndOf
 
 export default function Tokens() {
   const { data, error } = useUserBalanceAndOffers();
-  const { data: daoManagerData } = useGraphQL(getDaoManagerQuery);
+  const { data: daoManagerData } = useSubgraphGraphQL(getDaoManagerQuery);
 
   const totalVotingPower = bigIntToNum(daoManagerData?.daoManager?.totalVotingPower || BigInt(0));
   const userVotingPower = ((100 * (data?.balance?.votingPower || 0)) / totalVotingPower).toFixed(2);
