+++ hooks/useDelegationStatus.ts
@@ -1,7 +1,7 @@
 import { useAccount } from "wagmi";
 
 import { getDelegationUsers } from "@graphql/queries/subgraph/get-delegation-users-query";
-import { useGraphQL } from "@graphql/useGraphql";
+import { useSubgraphGraphQL } from "@graphql/subgraph";
 
 import { isSameAddress } from "@lib/utils";
 
@@ -12,7 +12,7 @@ const REFRESH_INTERVAL_MS = 10000;
 export default function useDelegationStatus() {
   const { address } = useAccount();
 
-  const { data, error, isLoading } = useGraphQL(address ? getDelegationUsers : null, {
+  const { data, error, isLoading } = useSubgraphGraphQL(address ? getDelegationUsers : null, {
     refreshInterval: REFRESH_INTERVAL_MS,
   });
