+++ pages/api/pdf/resolutions/[id].ts
@@ -6,11 +6,11 @@ import { OdooUser, ResolutionEntity, ResolutionEntityEnhanced } from "types";
 
 import React from "react";
 
-import { fetcherGraphqlPublic, isLegacyClientEnabled, legacyFetcherGraphqlPublic } from "@graphql/client";
 import odooClient from "@graphql/odoo";
 import { getUsersQuery } from "@graphql/queries/get-users.query";
 import { getLegacyResolutionQuery } from "@graphql/queries/subgraph/get-legacy-resolution-query";
 import { getResolutionQuery } from "@graphql/queries/subgraph/get-resolution-query";
+import { fetcherGraphqlPublic, isLegacyClientEnabled, legacyFetcherGraphqlPublic } from "@graphql/subgraph";
 
 import { getEnhancedResolutionMapper } from "@lib/resolutions/common";
 import { sessionOptions } from "@lib/session";
